remove deprecated dependencies
