+++ components/Post/Photo/index.ts
@@ -1,0 +1,3 @@
+import { Photo } from './Photo'
+
+export default Photo
