+++ package.json
@@ -11,6 +11,8 @@
   "dependencies": {
     "@grantcodes/themer": "^2.0.0",
     "@postr/core": "^1.10.1",
+    "@stripe/react-stripe-js": "^2.1.1",
+    "@stripe/stripe-js": "^1.54.1",
     "classnames": "^2.3.2",
     "dotenv": "^16.3.1",
     "eva-icons": "^1.1.3",
@@ -21,19 +23,18 @@
     "next": "^13.4.9",
     "next-fonts": "^1.5.1",
     "node-fetch": "^3.3.1",
+    "photoswipe": "^5.3.8",
     "pigeon-maps": "^0.21.3",
     "pigeon-overlay": "^0.2.3",
     "pluralize": "^8.0.0",
     "pouchdb-adapter-http": "^7.2.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-image-lightbox": "^5.1.4",
     "react-leaflet": "^4.2.1",
-    "react-stripe-elements": "^6.1.2",
+    "react-photoswipe-gallery": "^2.2.7",
     "recharts": "^2.7.2",
     "sass": "^1.63.6",
     "sharp": "^0.32.1",
-    "stripe": "^12.12.0",
     "svg-inline-loader": "^0.8.2",
     "viewport-mercator-project": "^7.0.4"
   },
